+++ HISTORY.rst
@@ -2,7 +2,12 @@
 History
 =======
 
-0.3.0 (XXXX-XX-XX)
+0.3.1 (XXXX-XX-XX)
+------------------
+
+* Add Jackknife+-after-Bootstrap method and add mean and median as aggregation functions
+
+0.2.4 (XXXX-XX-XX)
 ------------------
 
 * Renaming estimators.py module to regression.py